+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file, following the suggestions of [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+- Add `auth/label_comp_id` to `ComponentExpression`
+
 ## [v1.3.0] - 2025-03-18
 
 - Support for pydantic v2