+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file, following the suggestions of [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [v1.4.0] - 2025-04-09
 
 - Add `auth/label_comp_id` to `ComponentExpression`
 - Add MVSJ to MVSX archive functionality